requirements.txt

argparse
whois
dnspython
requests
shodan
colorama